# .history/docs/api/openapi_20260204163400.yaml
openapi: 3.1.0
info:
  title: Discord Clone API
  version: 1.0.0
  description: |
    API REST pour le clone Discord.
    
    ## Authentification
    Toutes les routes (sauf `/auth/register` et `/auth/login`) nécessitent un token JWT.
    
    ```
    Authorization: Bearer <access_token>
    ```
  contact:
    name: Discord Clone Team
    email: team@discord-clone.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://api.discord-clone.example.com/api/v1
    description: Production
...
paths:
...
  /servers:
    get:
      tags: [Servers]
      summary: List user's servers
      operationId: listServers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: List of servers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Servers]
      summary: Create a new server
      operationId: createServer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServerRequest'
      responses:
        '201':
          description: Server created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /servers/{serverId}:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Servers]
      summary: Get server details
      operationId: getServer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Server details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerWithChannels'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Servers]
      summary: Update server
      operationId: updateServer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
      responses:
        '200':
          description: Server updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Servers]
      summary: Delete server
      operationId: deleteServer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Server deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /servers/{serverId}/channels:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Channels]
      summary: List channels in server
      operationId: listChannels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Channels]
      summary: Create a channel
      operationId: createChannel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /servers/{serverId}/members:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Members]
      summary: List server members
      operationId: listMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /servers/{serverId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/ServerId'
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Members]
      summary: Get member details
      operationId: getMember
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Members]
      summary: Update member role
      operationId: updateMemberRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Member role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Members]
      summary: Kick member
      operationId: kickMember
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Member kicked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /servers/{serverId}/invites:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Invites]
      summary: List server invites
      operationId: listInvites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of invites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Invites]
      summary: Create invite
      operationId: createInvite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteWithUrl'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Role:
      type: string
      enum: [Owner, Admin, Moderator, Member]
      description: "Member role in PascalCase (matches Rust enum serialization)"
    ChannelKind:
      type: string
      enum: [Text]
      description: "Channel kind - currently only 'Text' is implemented (Voice planned)"
    Server:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        icon_url:
          type: string
          nullable: true
        owner_id:
          type: string
          format: uuid
        member_count:
          type: integer
        my_role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
    ServerWithChannels:
      allOf:
        - $ref: '#/components/schemas/Server'
        - type: object
          properties:
            channels:
              type: array
              items:
                $ref: '#/components/schemas/ChannelSummary'
    ChannelSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        kind:
          $ref: '#/components/schemas/ChannelKind'
    ServerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Server'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateServerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          maxLength: 1000
    UpdateServerRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        icon_url:
          type: string
          format: uri
    Channel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        server_id:
          type: string
          format: uuid
        name:
          type: string
        kind:
          $ref: '#/components/schemas/ChannelKind'
        created_at:
          type: string
          format: date-time
    ChannelList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
    CreateChannelRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        kind:
          $ref: '#/components/schemas/ChannelKind'
          default: Text
...
    Member:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        server_id:
          type: string
          format: uuid
        username:
          type: string
        avatar_url:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
        joined_at:
          type: string
          format: date-time
    MemberDetails:
      allOf:
        - $ref: '#/components/schemas/Member'
        - type: object
          properties:
            email:
              type: string
              format: email
            permissions:
              type: array
              items:
                type: string
    MemberList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    UpdateRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, moderator, member]
    Invite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        server_id:
          type: string
          format: uuid
        created_by:
          type: object
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
        uses:
          type: integer
        max_uses:
          type: integer
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    InviteWithUrl:
      allOf:
        - $ref: '#/components/schemas/Invite'
        - type: object
          properties:
            invite_url:
              type: string
              format: uri
...
    InviteList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invite'
    CreateInviteRequest:
      type: object
      properties:
        max_uses:
          type: integer
          minimum: 1
          nullable: true
        expires_in_hours:
          type: integer
          minimum: 1
          nullable: true
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer